+++ engine/src/org/pentaho/di/trans/steps/mapping/MappingMeta.java
@@ -2,7 +2,7 @@
  *
  * Pentaho Data Integration
  *
- * Copyright (C) 2002-2016 by Pentaho : http://www.pentaho.com
+ * Copyright (C) 2002-2017 by Pentaho : http://www.pentaho.com
  *
  *******************************************************************************
  *
@@ -56,10 +56,10 @@ import org.pentaho.di.resource.ResourceEntry;
 import org.pentaho.di.resource.ResourceEntry.ResourceType;
 import org.pentaho.di.resource.ResourceNamingInterface;
 import org.pentaho.di.resource.ResourceReference;
+import org.pentaho.di.trans.StepWithMappingMeta;
 import org.pentaho.di.trans.Trans;
 import org.pentaho.di.trans.TransMeta;
 import org.pentaho.di.trans.TransMeta.TransformationType;
-import org.pentaho.di.trans.step.BaseStepMeta;
 import org.pentaho.di.trans.step.StepDataInterface;
 import org.pentaho.di.trans.step.StepIOMeta;
 import org.pentaho.di.trans.step.StepIOMetaInterface;
@@ -82,13 +82,11 @@ import org.w3c.dom.Node;
  *
  */
 
-public class MappingMeta extends BaseStepMeta implements StepMetaInterface, HasRepositoryInterface {
-  private static Class<?> PKG = MappingMeta.class; // for i18n purposes, needed by Translator2!!
-  private String transName;
-  private String fileName;
-  private String directoryPath;
-  private ObjectId transObjectId;
-  private ObjectLocationSpecificationMethod specificationMethod;
+public class MappingMeta extends StepWithMappingMeta implements StepMetaInterface, HasRepositoryInterface {
+
+  static {
+    PKG = MappingMeta.class; // for i18n purposes, needed by Translator2!!
+  }
 
   private List<MappingIODefinition> inputMappings;
   private List<MappingIODefinition> outputMappings;
@@ -618,101 +616,6 @@ public class MappingMeta extends BaseStepMeta implements StepMetaInterface, HasR
     return loadMappingMeta( mappingMeta, rep, null, space );
   }
 
-  public static final synchronized TransMeta loadMappingMeta( MappingMeta mappingMeta, Repository rep,
-      IMetaStore metaStore, VariableSpace space ) throws KettleException {
-    return loadMappingMeta( mappingMeta, rep, metaStore, space, true );
-  }
-
-  public static final synchronized TransMeta loadMappingMeta( MappingMeta mappingMeta, Repository rep,
-      IMetaStore metaStore, VariableSpace space, boolean share ) throws KettleException {
-    TransMeta mappingTransMeta = null;
-
-    switch ( mappingMeta.getSpecificationMethod() ) {
-      case FILENAME:
-        String realFilename = space.environmentSubstitute( mappingMeta.getFileName() );
-        try {
-          // OK, load the meta-data from file...
-          //
-          // Don't set internal variables: they belong to the parent thread!
-          // PDI-3064 do not share with parent variable space
-          mappingTransMeta = new TransMeta( realFilename, metaStore, rep, true, null, null );
-          mappingTransMeta.getLogChannel().logDetailed(
-            "Loading Mapping from repository",
-            "Mapping transformation was loaded from XML file [" + realFilename + "]" );
-        } catch ( Exception e ) {
-          throw new KettleException( BaseMessages.getString( PKG, "MappingMeta.Exception.UnableToLoadMapping" ), e );
-        }
-        break;
-
-      case REPOSITORY_BY_NAME:
-        String realTransname = space.environmentSubstitute( mappingMeta.getTransName() );
-        String realDirectory = space.environmentSubstitute( mappingMeta.getDirectoryPath() );
-
-        if ( rep == null ) { // hardening because TransMeta.setRepositoryOnMappingSteps(); might be missing in special
-                             // situations
-          throw new KettleException( BaseMessages.getString(
-            PKG, "MappingMeta.Exception.InternalErrorRepository.Message" ) );
-        }
-
-        if ( !Utils.isEmpty( realTransname ) && !Utils.isEmpty( realDirectory ) && rep != null ) {
-          RepositoryDirectoryInterface repdir = rep.findDirectory( realDirectory );
-          if ( repdir != null ) {
-            try {
-              // reads the last revision in the repository...
-              //
-              mappingTransMeta = rep.loadTransformation( realTransname, repdir, null, true, null ); // TODO: FIXME:
-                                                                                                    // Should we pass in
-                                                                                                    // external
-                                                                                                    // MetaStore into
-                                                                                                    // Repository
-                                                                                                    // methods?
-
-              mappingTransMeta.getLogChannel().logDetailed(
-                "Loading Mapping from repository",
-                "Mapping transformation [" + realTransname + "] was loaded from the repository" );
-            } catch ( Exception e ) {
-              throw new KettleException( "Unable to load transformation [" + realTransname + "]", e );
-            }
-          } else {
-            throw new KettleException( BaseMessages.getString(
-              PKG, "MappingMeta.Exception.UnableToLoadTransformation", realTransname )
-              + realDirectory );
-          }
-        } else {
-          throw new KettleException( BaseMessages.getString(
-            PKG, "MappingMeta.Exception.UnableToLoadTransformationNameOrDirNotGiven" ) );
-        }
-        break;
-
-      case REPOSITORY_BY_REFERENCE:
-        // Read the last revision by reference...
-        if ( rep == null ) { // hardening because TransMeta.setRepositoryOnMappingSteps(); might be missing in special
-                             // situations
-          throw new KettleException( BaseMessages.getString(
-            PKG, "MappingMeta.Exception.InternalErrorRepository.Message" ) );
-        }
-        mappingTransMeta = rep.loadTransformation( mappingMeta.getTransObjectId(), null );
-        break;
-      default:
-        break;
-    }
-
-    // Pass some important information to the mapping transformation metadata:
-    //
-    if ( mappingTransMeta == null ) { // hardening because TransMeta might have issues in special situations
-      throw new KettleException( BaseMessages.getString(
-        PKG, "MappingMeta.Exception.InternalErrorTransMetaIsNULL.Message" ) );
-    }
-    if ( share ) {
-      mappingTransMeta.copyVariablesFrom( space );
-    }
-    mappingTransMeta.setRepository( rep );
-    mappingTransMeta.setMetaStore( metaStore );
-    mappingTransMeta.setFilename( mappingTransMeta.getFilename() );
-
-    return mappingTransMeta;
-  }
-
   public void check( List<CheckResultInterface> remarks, TransMeta transMeta, StepMeta stepMeta,
     RowMetaInterface prev, String[] input, String[] output, RowMetaInterface info, VariableSpace space,
     Repository repository, IMetaStore metaStore ) {
@@ -800,51 +703,6 @@ public class MappingMeta extends BaseStepMeta implements StepMetaInterface, HasR
 
   public StepDataInterface getStepData() {
     return new MappingData();
-  }
-
-  /**
-   * @return the directoryPath
-   */
-  public String getDirectoryPath() {
-    return directoryPath;
-  }
-
-  /**
-   * @param directoryPath
-   *          the directoryPath to set
-   */
-  public void setDirectoryPath( String directoryPath ) {
-    this.directoryPath = directoryPath;
-  }
-
-  /**
-   * @return the fileName
-   */
-  public String getFileName() {
-    return fileName;
-  }
-
-  /**
-   * @param fileName
-   *          the fileName to set
-   */
-  public void setFileName( String fileName ) {
-    this.fileName = fileName;
-  }
-
-  /**
-   * @return the transName
-   */
-  public String getTransName() {
-    return transName;
-  }
-
-  /**
-   * @param transName
-   *          the transName to set
-   */
-  public void setTransName( String transName ) {
-    this.transName = transName;
   }
 
   /**
@@ -976,21 +834,6 @@ public class MappingMeta extends BaseStepMeta implements StepMetaInterface, HasR
    */
   public void setRepository( Repository repository ) {
     this.repository = repository;
-  }
-
-  /**
-   * @return the transObjectId
-   */
-  public ObjectId getTransObjectId() {
-    return transObjectId;
-  }
-
-  /**
-   * @param transObjectId
-   *          the transObjectId to set
-   */
-  public void setTransObjectId( ObjectId transObjectId ) {
-    this.transObjectId = transObjectId;
   }
 
   /**
@@ -1144,4 +987,5 @@ public class MappingMeta extends BaseStepMeta implements StepMetaInterface, HasR
   public void setMetaStore( IMetaStore metaStore ) {
     this.metaStore = metaStore;
   }
+
 }
